requirements.txt
biopython~=1.84